backend/requirements.txt
# FastAPI and Server
fastapi==0.110.1
uvicorn==0.25.0
python-multipart>=0.0.9

# Database and ORM
asyncpg>=0.29.0
databases>=0.8.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Authentication and Security  
pyjwt>=2.10.1
passlib>=1.7.4
bcrypt>=4.1.0
python-jose>=3.3.0
cryptography>=42.0.8

# Supabase
supabase>=2.9.0
postgrest>=0.16.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
requests>=2.31.0

# Development and Testing
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Additional libraries
pandas>=2.2.0
numpy>=1.26.0
tzdata>=2024.2
typer>=0.9.0

# WebSocket support
websockets>=12.0

# HTTP client
httpx>=0.27.0